+++ backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/endpoint/interfaces/UsersApi.java
@@ -117,7 +117,7 @@ public interface UsersApi {
         value = "/users/{id}"
     )
     default ResponseEntity<Void> usersIdDelete(
-        @Parameter(name = "id", description = "ID of the user that is retreived", required = true, schema = @Schema(description = "")) @PathVariable("id") Integer id
+        @Parameter(name = "id", description = "ID of the user that is retreived", required = true, schema = @Schema(description = "")) @PathVariable("id") Long id
     ) {
         return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
 
@@ -153,7 +153,7 @@ public interface UsersApi {
         produces = { "application/json" }
     )
     default ResponseEntity<UserDto> usersIdGet(
-        @Parameter(name = "id", description = "ID of the user that is retreived", required = true, schema = @Schema(description = "")) @PathVariable("id") Integer id
+        @Parameter(name = "id", description = "ID of the user that is retreived", required = true, schema = @Schema(description = "")) @PathVariable("id") Long id
     ) {
         getRequest().ifPresent(request -> {
             for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
@@ -199,7 +199,7 @@ public interface UsersApi {
         consumes = { "application/json" }
     )
     default ResponseEntity<Void> usersIdPut(
-        @Parameter(name = "id", description = "ID of the user that is retreived", required = true, schema = @Schema(description = "")) @PathVariable("id") Integer id,
+        @Parameter(name = "id", description = "ID of the user that is retreived", required = true, schema = @Schema(description = "")) @PathVariable("id") Long id,
         @Parameter(name = "UserWithPasswordDto", description = "", required = true, schema = @Schema(description = "")) @Valid @RequestBody UserWithPasswordDto userWithPasswordDto
     ) {
         return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
